Rename HashShuffleReader to BlockStoreShuffleReader.
